pyproject.toml
[project]
name = "news-latter-embryo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "arxiv>=2.4.0",
    "feedparser>=6.0.12",
    "google-genai>=1.65.0",
    "google-generativeai>=0.8.6",
    "googlenewsdecoder>=0.1.7",
    "lxml-html-clean>=0.4.4",
    "newspaper4k[nlp]>=0.9.5",
    "python-dotenv>=1.2.2",
    "tenacity>=9.1.4",
    "networkx>=3.3",
    "matplotlib>=3.9",
    "koreanize-matplotlib>=0.1.1",
    "setuptools>=82.0.1",
]